I1

+++ 1float.cpp
@@ -4,11 +4,11 @@
 #include <chrono>
 
 int n = 10000000;
-double pi = 3.14159265359;	
+float pi = 3.14159265359;	
 
 int main() {
-	double* arr = (double*)malloc(n * sizeof(double));
-	double sum = 0;
+	float* arr = (float*)malloc(n * sizeof(double));
+	float sum = 0;
 
 #pragma acc enter data create(arr[0:n],sum)
 
@@ -21,7 +21,7 @@ int main() {
 #pragma acc parallel loop present(arr[0:n],sum) reduction(+:sum)
 	for (int i = 0; i < n; i++) sum += arr[i];
 #pragma acc exit data delete(arr[0:n]) copyout(sum)
-	printf("%0.30lf\n", sum);
+	printf("%0.30f\n", sum);
 
 	free(arr);
 	return 0;
